Implemented twitter deep-link authentication
