+++ lib/screens/auth/base/sign_in.dart
@@ -1,18 +1,68 @@
+import 'dart:async';
+
+import 'package:delphis_app/bloc/auth/auth_bloc.dart';
+import 'package:delphis_app/constants.dart';
 import 'package:delphis_app/design/colors.dart';
 import 'package:delphis_app/design/sizes.dart';
 import 'package:delphis_app/design/text_theme.dart';
 import 'package:delphis_app/screens/auth/base/widgets/loginWithTwitterButton/twitter_button.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:intl/intl.dart';
+import 'package:uni_links/uni_links.dart';
 import 'package:url_launcher/url_launcher.dart';
 
-class SignInScreen extends StatelessWidget {
+class SignInScreen extends StatefulWidget {
+  @override
+  _SignInScreenState createState() => _SignInScreenState();
+}
+
+class _SignInScreenState extends State<SignInScreen> {
   static const twitterWidgetHeight = 76.0;
   static const feedbackLink =
       'mailto:ned@chatham.ai?subject=Twitter%20Makes%20Me%20Mad';
+  StreamSubscription _deepLinkSubscription;
+  AuthBloc authBloc;
 
+  @override
+  void dispose() {
+    _deepLinkSubscription?.cancel();
+    super.dispose();
+  }
+  @override
+  void initState() {
+    super.initState();
+    
+    this.authBloc = BlocProvider.of<AuthBloc>(context);
+    this._deepLinkSubscription = getLinksStream().listen((String link) {
+      if (link.startsWith(Constants.twitterRedirectURLPrefix)) {
+            String token = RegExp("\\?dc=(.*)").firstMatch(link)?.group(1) ?? null;
+            if(token != null) {
+              this.authBloc.add(LoadedAuthEvent(token, true, false));
+              this.successfulLogin();
+            }
+          }
+    }, onError: (err) {
+      throw err;
+    });
+  }
+
+  void successfulLogin() {
+    Navigator.of(context)
+        .pushNamedAndRemoveUntil('/Home', (Route<dynamic> route) => false);
+  }
+  
+  void openLoginDialog() async {
+    var url = Constants.twitterLoginURL;
+    if (await canLaunch(url)) {
+      await launch(url);
+    } else {
+      throw 'Could not launch $url';
+    }
+  }
+  
   @override
   Widget build(BuildContext context) {
     return SafeArea(
@@ -55,8 +105,7 @@ class SignInScreen extends StatelessWidget {
                                 textAlign: TextAlign.center),
                             SizedBox(height: SpacingValues.mediumLarge),
                             LoginWithTwitterButton(
-                              onPressed: () => Navigator.of(context)
-                                  .pushNamed('/Auth/Twitter'),
+                              onPressed: () => openLoginDialog(),
                               width: constraints.maxWidth,
                               height: 56.0,
                             ),
@@ -82,4 +131,5 @@ class SignInScreen extends StatelessWidget {
       ),
     );
   }
+
 }
