--- lib/screens/auth/index.dart
@@ -1,1 +1,0 @@
-export 'auth.dart';
